Compare downloaded to files available from ASF
The script compares the contents of doownloaded folder against the list of available files from ASF. The list from ASF is obtained with API query. The script logs how many files are (i) already downloaded, (ii) missing, still need to be downloaded, and (iii) the files that are surplus are moved to a specified folder. The reason for surplus files was that we accidentally downloaded files for two different regions into the same folder.

Added is also the download-all.py script from ASF, that is used together with the CSV file of the missing products, to download them.  Run from terminal with csv file as the only argument.